+++ src/pages/{StoreCollection.slug}/[...].tsx
@@ -1,4 +1,4 @@
-import React, { Suspense, useMemo } from 'react'
+import React, { lazy, Suspense, SuspenseList, useMemo } from 'react'
 import Layout from 'src/views/Layout'
 import View, { Preview } from 'src/views/collection'
 import { graphql } from 'gatsby'
@@ -15,6 +15,14 @@ export type Props = PageProps<
   BrowserCollectionPageQueryQueryVariables & { slug: string }
 >
 
+const CollectionBanner = lazy(
+  () =>
+    import(
+      /* webpackMode: "eager" */
+      'src/components/sections/CollectionBanner'
+    )
+)
+
 const useSearchParams = ({ href }: Location) =>
   useMemo(() => href && parseSearchParamsState(new URL(href)), [href])
 
@@ -23,9 +31,24 @@ function Page(props: Props) {
 
   return (
     <Layout>
-      <Suspense fallback={<Preview {...props} />}>
-        {searchParams && <View {...props} searchParams={searchParams} />}
-      </Suspense>
+      <SuspenseList revealOrder="forwards">
+        <Suspense fallback={null}>
+          <CollectionBanner
+            image={{
+              desktop:
+                'https://fashioneurope.vtexassets.com/assets/vtex/assets-builder/fashioneurope.theme/2.7.0/images/search-banner___b133a2e011b0a025cdc7f9fb02645848.jpg',
+              mobile:
+                'https://fashioneurope.vtexassets.com/assets/vtex/assets-builder/fashioneurope.theme/2.7.0/images/search-banner___b133a2e011b0a025cdc7f9fb02645848.jpg',
+              alt: 'Collection Image',
+            }}
+            title={props.data.storeCollection?.seo.title ?? 'Collection'}
+            description="explore the collection"
+          />
+        </Suspense>
+        <Suspense fallback={<Preview />}>
+          {searchParams && <View {...props} searchParams={searchParams} />}
+        </Suspense>
+      </SuspenseList>
     </Layout>
   )
 }
