+++ src/pages/{StoreCollection.slug}/index.tsx
@@ -1,4 +1,4 @@
-import React, { Suspense, useMemo } from 'react'
+import React, { lazy, Suspense, SuspenseList, useMemo } from 'react'
 import Layout from 'src/views/Layout'
 import View, { Preview } from 'src/views/collection'
 import { graphql } from 'gatsby'
@@ -10,6 +10,14 @@ import type {
   CollectionPageQueryQuery,
   CollectionPageQueryQueryVariables,
 } from './__generated__/CollectionPageQuery.graphql'
+
+const CollectionBanner = lazy(
+  () =>
+    import(
+      /* webpackMode: "eager" */
+      'src/components/sections/CollectionBanner'
+    )
+)
 
 export type Props = PageProps<
   CollectionPageQueryQuery,
@@ -55,9 +63,24 @@ function Page(props: Props) {
 
   return (
     <Layout>
-      <Suspense fallback={<Preview {...props} />}>
-        <View {...props} searchParams={searchParams} />
-      </Suspense>
+      <SuspenseList revealOrder="forwards">
+        <Suspense fallback={null}>
+          <CollectionBanner
+            image={{
+              desktop:
+                'https://fashioneurope.vtexassets.com/assets/vtex/assets-builder/fashioneurope.theme/2.7.0/images/search-banner___b133a2e011b0a025cdc7f9fb02645848.jpg',
+              mobile:
+                'https://fashioneurope.vtexassets.com/assets/vtex/assets-builder/fashioneurope.theme/2.7.0/images/search-banner___b133a2e011b0a025cdc7f9fb02645848.jpg',
+              alt: 'Collection Image',
+            }}
+            title={props.data.storeCollection?.seo.title ?? 'Collection'}
+            description="explore the collection"
+          />
+        </Suspense>
+        <Suspense fallback={<Preview />}>
+          <View {...props} searchParams={searchParams} />
+        </Suspense>
+      </SuspenseList>
     </Layout>
   )
 }
