+++ src/views/collection/index.tsx
@@ -23,14 +23,6 @@ const ProductGallery = lazy(
     )
 )
 
-const CollectionBanner = lazy(
-  () =>
-    import(
-      /* webpackMode: "eager" */
-      'src/components/sections/CollectionBanner'
-    )
-)
-
 const SearchFilters = lazy(
   () =>
     import(
@@ -85,20 +77,6 @@ function View(props: Props) {
 
         {/* UI components */}
         <Suspense fallback={null}>
-          <CollectionBanner
-            image={{
-              desktop:
-                'https://fashioneurope.vtexassets.com/assets/vtex/assets-builder/fashioneurope.theme/2.7.0/images/search-banner___b133a2e011b0a025cdc7f9fb02645848.jpg',
-              mobile:
-                'https://fashioneurope.vtexassets.com/assets/vtex/assets-builder/fashioneurope.theme/2.7.0/images/search-banner___b133a2e011b0a025cdc7f9fb02645848.jpg',
-              alt: 'Collection Image',
-            }}
-            title={storeCollection?.seo.title ?? 'Collection'}
-            description="explore the collection"
-          />
-        </Suspense>
-
-        <Suspense fallback={null}>
           <SearchFilters facets={facets!.facets as any} />
         </Suspense>
 
@@ -113,26 +91,11 @@ function View(props: Props) {
   )
 }
 
-export function Preview(props: Omit<Props, 'searchParams'>) {
+export function Preview() {
   return (
-    <>
-      <Suspense fallback={null}>
-        <CollectionBanner
-          image={{
-            desktop:
-              'https://fashioneurope.vtexassets.com/assets/vtex/assets-builder/fashioneurope.theme/2.7.0/images/search-banner___b133a2e011b0a025cdc7f9fb02645848.jpg',
-            mobile:
-              'https://fashioneurope.vtexassets.com/assets/vtex/assets-builder/fashioneurope.theme/2.7.0/images/search-banner___b133a2e011b0a025cdc7f9fb02645848.jpg',
-            alt: 'Collection Image',
-          }}
-          title={props.data.storeCollection?.seo.title ?? 'Collection'}
-          description="explore the collection"
-        />
-      </Suspense>
-      <div className="h-96 flex center-items justify-center">
-        <Spinner />
-      </div>
-    </>
+    <div className="h-96 flex center-items justify-center">
+      <Spinner />
+    </div>
   )
 }
